+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
-## [0.2.0] - 2024-04-30
+## [0.2.0] - 2024-06-05
 
 ### Added
 - Support retrieving secrets definition from pod annotations (CNJR-4099)
 - Added support for configurable log levels using the `LOG_LEVEL` environment
   variable (CNJR-3733)
+- Added support for JWT authenticator field `token-app-property`, which makes
+  the `identity` configuration attribute optional (CNJR-4607)
 
 ### Changed
 - Updated Alpine base image to 3.19.1 (CONJSE-1852)